Buffer tests and (C, D) LQR parameters
Example 1 of  https://ieeexplore.ieee.org/document/8618689 works, and involves parameters C = [I; 0] and D = [0; I] (like in LQR). This example has a case where A is stable to begin with.

This same choice of parameters (C, D) fail when applied to a the test system p2p_clean_cont.m in which A is unstable.